♻️ rfac(custom): select project 코드 위치 수정
