+++ src/app/review/_components/current-week-review/CurrentWeekReviewContainer.tsx
@@ -1,9 +1,7 @@
 'use client';
 
-import { useAtom } from 'jotai';
+import { useAtomValue } from 'jotai';
 import { useEffect, useState } from 'react';
-
-import { DropdownItem } from '@/components/dropdown/Dropdown';
 
 import { highLightListAtom, lowLightListAtom } from '../../stores';
 import { ReviewType } from '../../types';
@@ -14,8 +12,8 @@ export const CurrentWeekReviewContainer = ({
 }: {
   category: ReviewType;
 }) => {
-  const [highLightList, setHighLightList] = useAtom(highLightListAtom);
-  const [lowLightList, setLowLightList] = useAtom(lowLightListAtom);
+  const highLightList = useAtomValue(highLightListAtom);
+  const lowLightList = useAtomValue(lowLightListAtom);
 
   const [reviewList, setReviewList] = useState(
     category === 'highLight' ? highLightList : lowLightList,
@@ -29,27 +27,6 @@ export const CurrentWeekReviewContainer = ({
     }
   }, [highLightList, lowLightList, category]);
 
-  const selectProject = (item: DropdownItem, reviewId: string | number) => {
-    const selectedItem = {
-      id: item.id,
-      content: item.name,
-      progressRate: 0,
-    };
-    const setter =
-      category === 'highLight' ? setHighLightList : setLowLightList;
-
-    setter((prev) =>
-      prev.map((review) =>
-        String(review.id) === String(reviewId)
-          ? {
-              ...review,
-              project: selectedItem,
-            }
-          : review,
-      ),
-    );
-  };
-
   return (
     <div className="w-full p-5 bg-surface-foreground rounded-xl flex flex-col gap-6 mb-3">
       {reviewList.map((el, index) => (
@@ -57,7 +34,6 @@ export const CurrentWeekReviewContainer = ({
           key={String(el.id)}
           category={category}
           index={index}
-          onSelect={(item) => selectProject(item, el.id)}
           {...el}
         />
       ))}
